refactor tests, extract data builder
