+++ NuKeeper.Tests/NuGet/Api/ApiPackageLookupTests.cs
@@ -2,7 +2,6 @@
 using System.Threading.Tasks;
 using NSubstitute;
 using NuGet.Packaging.Core;
-using NuGet.Protocol.Core.Types;
 using NuGet.Versioning;
 using NuKeeper.NuGet.Api;
 using NUnit.Framework;
@@ -28,7 +27,7 @@ namespace NuKeeper.Tests.NuGet.Api
         {
             var resultPackages = new List<PackageSearchMedatadataWithSource>
             {
-                BuildMetadata("TestPackage", 2, 3, 4)
+                PackageVersionTestData.BuildMetadata("TestPackage", 2, 3, 4)
             };
 
             var allVersionsLookup = MockVersionLookup(resultPackages);
@@ -48,7 +47,7 @@ namespace NuKeeper.Tests.NuGet.Api
             int expectedMajor, int expectedMinor, int expectedPatch)
         {
             var expectedUpdate = new NuGetVersion(expectedMajor, expectedMinor, expectedPatch);
-            var resultPackages = VersionsFor(dataRange);
+            var resultPackages = PackageVersionTestData.VersionsFor(dataRange);
             var allVersionsLookup = MockVersionLookup(resultPackages);
 
             IApiPackageLookup lookup = new ApiPackageLookup(allVersionsLookup);
@@ -67,7 +66,7 @@ namespace NuKeeper.Tests.NuGet.Api
             int expectedMajor, int expectedMinor, int expectedPatch)
         {
             var expectedUpdate = new NuGetVersion(expectedMajor, expectedMinor, expectedPatch);
-            var resultPackages = VersionsFor(dataRange);
+            var resultPackages = PackageVersionTestData.VersionsFor(dataRange);
             var allVersionsLookup = MockVersionLookup(resultPackages);
 
             IApiPackageLookup lookup = new ApiPackageLookup(allVersionsLookup);
@@ -86,7 +85,7 @@ namespace NuKeeper.Tests.NuGet.Api
             int expectedMajor, int expectedMinor, int expectedPatch)
         {
             var expectedUpdate = new NuGetVersion(expectedMajor, expectedMinor, expectedPatch);
-            var resultPackages = VersionsFor(dataRange);
+            var resultPackages = PackageVersionTestData.VersionsFor(dataRange);
             var allVersionsLookup = MockVersionLookup(resultPackages);
 
             IApiPackageLookup lookup = new ApiPackageLookup(allVersionsLookup);
@@ -98,101 +97,12 @@ namespace NuKeeper.Tests.NuGet.Api
             Assert.That(package.Identity.Version, Is.EqualTo(expectedUpdate));
         }
 
-        private static List<PackageSearchMedatadataWithSource> VersionsFor(VersionChange change)
-        {
-            switch (change)
-            {
-                case VersionChange.Major:
-                    return AllKindsOfVersions();
-
-                case VersionChange.Minor:
-                    return MinorVersions();
-
-                case VersionChange.Patch:
-                    return PatchVersions();
-
-                default:
-                    return new List<PackageSearchMedatadataWithSource>();
-            }
-        }
-
-
-        private static List<PackageSearchMedatadataWithSource> AllKindsOfVersions()
-        {
-            return new List<PackageSearchMedatadataWithSource>
-            {
-                BuildMetadata("TestPackage", 2, 3, 4),
-                BuildMetadata("TestPackage", 2, 1, 1),
-                BuildMetadata("TestPackage", 2, 0, 0),
-
-                BuildMetadata("TestPackage", 1, 3, 1),
-                BuildMetadata("TestPackage", 1, 3, 0),
-
-                BuildMetadata("TestPackage", 1, 2, 5),
-                BuildMetadata("TestPackage", 1, 2, 4),
-                BuildMetadata("TestPackage", 1, 2, 3),
-                BuildMetadata("TestPackage", 1, 2, 2),
-                BuildMetadata("TestPackage", 1, 2, 1),
-
-                BuildMetadata("TestPackage", 1, 1, 0),
-                BuildMetadata("TestPackage", 1, 0, 0)
-            };
-        }
-
-        private static List<PackageSearchMedatadataWithSource> MinorVersions()
-        {
-            return new List<PackageSearchMedatadataWithSource>
-            {
-                BuildMetadata("TestPackage", 1, 3, 1),
-                BuildMetadata("TestPackage", 1, 3, 0),
-
-                BuildMetadata("TestPackage", 1, 2, 5),
-                BuildMetadata("TestPackage", 1, 2, 4),
-                BuildMetadata("TestPackage", 1, 2, 3),
-                BuildMetadata("TestPackage", 1, 2, 2),
-                BuildMetadata("TestPackage", 1, 2, 1),
-
-                BuildMetadata("TestPackage", 1, 1, 0),
-                BuildMetadata("TestPackage", 1, 0, 0)
-            };
-        }
-
-        private static List<PackageSearchMedatadataWithSource> PatchVersions()
-        {
-            return new List<PackageSearchMedatadataWithSource>
-            {
-                BuildMetadata("TestPackage", 1, 2, 5),
-                BuildMetadata("TestPackage", 1, 2, 4),
-                BuildMetadata("TestPackage", 1, 2, 3),
-                BuildMetadata("TestPackage", 1, 2, 2),
-                BuildMetadata("TestPackage", 1, 2, 1),
-
-                BuildMetadata("TestPackage", 1, 1, 0),
-                BuildMetadata("TestPackage", 1, 0, 0)
-            };
-        }
-
         private static IPackageVersionsLookup MockVersionLookup(List<PackageSearchMedatadataWithSource> actualResults)
         {
             var allVersions = Substitute.For<IPackageVersionsLookup>();
             allVersions.Lookup(Arg.Any<string>())
                 .Returns(actualResults);
             return allVersions;
-        }
-
-        private static PackageSearchMedatadataWithSource BuildMetadata(string source, int major, int minor, int patch)
-        {
-            var version = new NuGetVersion(major, minor, patch);
-            var metadata = MetadataWithVersion(source, version);
-            return new PackageSearchMedatadataWithSource(source, metadata);
-        }
-
-        private static IPackageSearchMetadata MetadataWithVersion(string id, NuGetVersion version)
-        {
-            var metadata = Substitute.For<IPackageSearchMetadata>();
-            var identity = new PackageIdentity(id, version);
-            metadata.Identity.Returns(identity);
-            return metadata;
         }
 
         private PackageIdentity CurrentVersion123(string packageId)
